refactor(terraform): remove dead legacy sequential executors
Deletes internal/exec/terraform_executor.go and
terraform_affected_graph.go, plus the orphaned processTerraformComponent/
executeTerraformAffectedComponentInDepOrder/executeAffectedComponents chain
in terraform_utils.go and terraform_affected.go. Confirmed via call-graph
grep: none of these have had a live (non-test) caller since the terraform
scheduler migration — every --all/--affected/--query path already routes
through pkg/scheduler/adapters instead. extractAffectedNodeIDs is relocated
(it's still used by the live ExecuteTerraformAffectedWithContext).

Split out from the hooks bulk-dispatch fix in this same PR so the behavior
change stays independently bisectable from this pure deletion.
