+++ internal/exec/terraform_utils.go
@@ -1,10 +1,8 @@
 package exec
 
 import (
-	"bytes"
 	"context"
 	"errors"
-	"fmt"
 	"os"
 	"path/filepath"
 	"slices"
@@ -18,20 +16,10 @@ import (
 	"github.com/cloudposse/atmos/pkg/perf"
 	"github.com/cloudposse/atmos/pkg/schema"
 	tfoutput "github.com/cloudposse/atmos/pkg/terraform/output"
-	"github.com/cloudposse/atmos/pkg/ui"
 	u "github.com/cloudposse/atmos/pkg/utils"
 )
 
 const commandStr = "command"
-
-// affectedDepOrderParams holds parameters for executing an affected component in dependency order.
-type affectedDepOrderParams struct {
-	AffectedComponent string
-	AffectedStack     string
-	ParentComponent   string
-	ParentStack       string
-	Dependents        []schema.Dependent
-}
 
 // shouldProcessStacks determines whether to process stacks and check stack configuration.
 // Based on the command type and provided arguments.
@@ -386,164 +374,6 @@ func mirrorTargetIncluded(atmosConfig *schema.AtmosConfiguration, componentName 
 	}
 	passed, ok := queryResult.(bool)
 	return ok && passed, nil
-}
-
-// processTerraformComponent performs filtering and execution logic for a single Terraform component.
-// Returns true if the component was processed (passed all filters), false otherwise.
-// The executeFn parameter allows callers to inject a custom executor (used for testing without gomonkey).
-func processTerraformComponent(
-	atmosConfig *schema.AtmosConfiguration,
-	info *schema.ConfigAndStacksInfo,
-	stackName, componentName string,
-	componentSection map[string]any,
-	logFunc func(msg any, keyvals ...any),
-	executeFn func(schema.ConfigAndStacksInfo, ...ShellCommandOption) error,
-) (bool, error) {
-	metadataSection, ok := componentSection[cfg.MetadataSectionName].(map[string]any)
-	if !ok {
-		return false, nil
-	}
-
-	// Skip abstract components.
-	if metadataType, ok := metadataSection["type"].(string); ok && metadataType == "abstract" {
-		return false, nil
-	}
-
-	// Skip disabled components.
-	if !isComponentEnabled(metadataSection, componentName) {
-		return false, nil
-	}
-
-	command := fmt.Sprintf("atmos terraform %s %s -s %s", info.SubCommand, componentName, stackName)
-
-	if info.Query != "" {
-		queryResult, err := u.EvaluateYqExpression(atmosConfig, componentSection, info.Query)
-		if err != nil {
-			return false, err
-		}
-
-		if queryPassed, ok := queryResult.(bool); !ok || !queryPassed {
-			logFunc("Skipping the component because the query criteria not satisfied", commandStr, command, "query", info.Query)
-			return false, nil
-		}
-	}
-
-	logFunc("Executing", commandStr, command)
-
-	// Show user-facing progress for dry-run mode.
-	if info.DryRun {
-		ui.Successf("Would %s `%s` in `%s` (dry run)", info.SubCommand, componentName, stackName)
-		return true, nil
-	}
-
-	info.Component = componentName
-	info.ComponentFromArg = componentName
-	info.Stack = stackName
-	info.StackFromArg = stackName
-
-	// When a per-component hook is registered, capture this component's output
-	// and invoke the hook immediately after execution so each component receives
-	// its own CI summary entry rather than sharing the final global call.
-	if info.PerComponentHook != nil {
-		var stdoutBuf, stderrBuf bytes.Buffer
-		execErr := executeFn(*info, WithStdoutCapture(&stdoutBuf), WithStderrCapture(&stderrBuf))
-		combined := stdoutBuf.String()
-		if s := stderrBuf.String(); s != "" {
-			combined += "\n" + s
-		}
-		compInfo := *info // snapshot with this component's Component/Stack values set.
-		info.PerComponentHook(&compInfo, combined, execErr)
-		return true, execErr
-	}
-
-	if err := executeFn(*info); err != nil {
-		return true, err
-	}
-
-	return true, nil
-}
-
-// logFuncForDryRun returns log.Info for dry-run mode, log.Debug otherwise.
-func logFuncForDryRun(dryRun bool) func(msg any, keyvals ...any) {
-	if dryRun {
-		return log.Info
-	}
-	return log.Debug
-}
-
-// shouldProcessDependent checks if a dependent should be processed.
-func shouldProcessDependent(dep *schema.Dependent, affectedList []schema.Affected, includeDependents bool) bool {
-	if dep.IncludedInDependents {
-		return false
-	}
-	return includeDependents || isComponentInStackAffected(affectedList, dep.StackSlug)
-}
-
-// isNonTerraformDependent returns true when dep is a dependent that
-// `atmos terraform` must skip during recursion: a helmfile or packer
-// component, or any other non-terraform type. Dependents with an empty
-// ComponentType are treated as terraform for backward compatibility. See
-// issue #2361.
-func isNonTerraformDependent(dep *schema.Dependent) bool {
-	return dep.ComponentType != "" && dep.ComponentType != cfg.TerraformComponentType
-}
-
-// executeTerraformAffectedComponentInDepOrder recursively processes the affected components in the dependency order.
-func executeTerraformAffectedComponentInDepOrder(
-	info *schema.ConfigAndStacksInfo,
-	affectedList []schema.Affected,
-	params *affectedDepOrderParams,
-	args *DescribeAffectedCmdArgs,
-) error {
-	logFunc := logFuncForDryRun(info.DryRun)
-
-	info.Component = params.AffectedComponent
-	info.ComponentFromArg = params.AffectedComponent
-	info.Stack = params.AffectedStack
-
-	command := fmt.Sprintf("atmos terraform %s %s -s %s", info.SubCommand, params.AffectedComponent, params.AffectedStack)
-
-	if args.IncludeDependents && params.ParentComponent != "" && params.ParentStack != "" {
-		logFunc("Executing", commandStr, command, "dependency of component", params.ParentComponent, "in stack", params.ParentStack)
-	} else {
-		logFunc("Executing", commandStr, command)
-	}
-
-	if !info.DryRun {
-		if err := ExecuteTerraform(*info); err != nil {
-			return err
-		}
-	}
-
-	for i := 0; i < len(params.Dependents); i++ {
-		dep := &params.Dependents[i]
-		// Defense-in-depth: when --include-dependents is set, the dependency graph
-		// may include helmfile/packer dependents. `atmos terraform` must skip those
-		// — they belong to their own subcommands. See issue #2361.
-		if isNonTerraformDependent(dep) {
-			continue
-		}
-		if !shouldProcessDependent(dep, affectedList, args.IncludeDependents) {
-			continue
-		}
-		err := executeTerraformAffectedComponentInDepOrder(
-			info,
-			affectedList,
-			&affectedDepOrderParams{
-				AffectedComponent: dep.Component,
-				AffectedStack:     dep.Stack,
-				ParentComponent:   params.AffectedComponent,
-				ParentStack:       params.AffectedStack,
-				Dependents:        dep.Dependents,
-			},
-			args,
-		)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
 }
 
 // parseUploadStatusFlag parses the upload status flag from the arguments.
